rocket_mall/requirements.txt
aiofiles==0.5.0
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-ckeditor==5.9.0
django-cors-headers==3.5.0
django-crontab==0.7.1
django-filter==2.3.0
django-haystack==3.0
django-js-asset==1.2.2
django-modelcluster==5.0.2
django-redis==4.12.1
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
draftjs-exporter==2.1.7
elasticsearch==2.4.1
email-validator==1.1.1
et-xmlfile==1.0.1
eventlet==0.26.1
fastapi==0.61.1
filelock==3.0.12
greenlet==0.4.16
h11==0.9.0
html5lib==1.1
httptools==0.1.1
idna==2.10
importlib-metadata==1.7.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
kombu==4.6.11
l18n==2018.5
MarkupSafe==1.1.1
monotonic==1.5
mutagen==1.45.1
oauthlib==3.1.0
openpyxl==3.0.4
pbr==5.4.5
Pillow==7.2.0
py3Fdfs==2.2.0
pycparser==2.20
pycryptodomex==3.9.4
pydantic==1.6.1
PyJWT==1.7.1
PyMySQL==0.10.0
pyOpenSSL==19.1.0
python-alipay-sdk==2.1.1
python3-openid==3.2.0
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
soupsieve==2.0.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
starlette==0.13.6
stevedore==3.2.0
tablib==2.0.0
twilio==6.44.2
Unidecode==1.1.1
urllib3==1.25.10
uvicorn==0.11.8
uvloop==0.14.0
vine==1.3.0
virtualenv==20.0.30
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
webencodings==0.5.1
websockets==8.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
zipp==3.1.0
uwsgi==2.0.18